feat(tmux): make tmux ready as a daily driver
Changed some tmux settings that makes is behave more like vim. Also
added a wrapper function which checks if there is a `.tmux` file in the
current directory, and if so, sources that file. This file creates a
session with a predefined window setup, which means that we don't have
to rely on our window manager to create multiple vim instances for
example. What's even cooler is that we can run commands in the created
splits, for example starting a build server.

See:
  - https://www.youtube.com/watch?v=ogeVqNOStQs