chore(dotfiles): rename config to .config and local to .local
This makes it easier when we create the install script to symlink all
files in the dotfiles directory to the home directory since we don't
have to append a `.` to the names. It also hide a lot of clutter in
CommandT when searching from the root directory since we must search
with a `.` for the files in `.config` and `.local` to show up.