feat(nvim): add loupe plugin
Loupe solves the problem of regex in vim by prepending `/\v`
(very magic) to searches where it believes it is wanted. It should
also solve the problem with seeing which word we are currently on, but
I am having trouble getting that to work. So it will be something I will
continue to work on. But Loupe is a big improvement to search anyways.