feat(vim): dip into using lua for configuration
Trying to use lua for configuration where it is possible. I have not
being interested in learning vimscript since I feel like it is quite
odd, but lua on the other hand seems quite nice and I am ready to dip
into it where I can. I will try to convert more and more of my config
into lua at some point, but at this point it is a bit limited due to
the fact that there is no lua implementation for autocommands.
Hopefully it should be fully supported in the 0.5 release if neovim.

I also switched to use `packer.nvim` as a package manager as it is
written and configured in lua. It was also getting a bit tedious to
manually update and manage plugins using git submodules. I could
probably have written a simple shell-script so solve the update part,
but I had all kinds of problems when I tried to reinstall submodules so
I felt like a plugin manager was the easiest solution.