feat(xorg): switch from dwm to awesome
Switching to awesome from dwm since i don't want to learn c in order to
customize it. Since awesome is configured in lua I now have another
incentive to learn it. In about an hour the feel was basically the same
as with dwm (or maybe even better?). This isn't that weird since awesome
started out as a fork of dwm. The only thing that I miss from my dwm
setup is the statusbar, but I will have to fix that at some point.