fix(vim): don't set spellfile
Setting `spellfile` leads neovim to think that that it should only look
for words in that file whereas it seems like it should do what I want
when I read the docs. However not setting `spellfile` fixes the problem
since neovim correctly places the added words in the spell directory.