feat(nvim): add lspsaga code action lighbulb
Display a lighbulb when a code action is available. We also lower the
`updatetime` option to `300` instead of the default `4000`. This makes
the lighbulb show up and disappear more quickly, which is nice. This
option does also effect when we write to swap ans such things, so we
might see some performance hiccups if it is happening to often. In that
case, we just have to increase the value a bit.