feat(dotfiles): move to nixos and nix
I am really intrigued by the fact that I can manage my whole system by
changing a couple of dotfiles, so nix seems to be the way to go. I still
like Arch and all the things that come with it (lightweight, AUR) etc.
but the ability to control my whole system with dotfiles seems really
good nice right now.