feat: replace awesomewm with i3
I could not get screen share to work on awesome reliably, and that's a
deal-breaker since it is needed when I work as a TA during Covid. The
only downside so far is that I cannot share a single window, just the
whole screen. But I can live with that, just have to make sure I don't
have anything sensitive on my screen :face_with_monocle: