feat(dots): start on dotfiles manager
Started hacking on a dotfiles manager that I will refer to as `dots`
for now. I want it to be self-contained in this repo, so that it is as
simple as cloning this repo and then run a command (no need to install
any dependencies). Decided to try out [bun](https://bun.sh/) for this,
and vendored the bun executable with the repo.

My intention is for the `dots` directory to be the root directory at
some point and I will therefor move more and more aspects to `dots`
while I add the functionality I want.

Credits: This is inspired by [Greg Hurrell's](https://github.com/wincent)
[fig](https://github.com/wincent/wincent/tree/main/fig):